fraud-detection-platform/requirements.txt
dbt-core==1.8.0
dbt-duckdb==1.8.0
duckdb==1.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
streamlit>=1.32.0
plotly>=5.19.0
mlflow>=2.10.0
faker>=24.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pydantic>=2.6.0
PyYAML>=6.0.0
joblib>=1.3.0